+++ lib/widgets/dialogs/filter_editors/cover_selection_dialog.dart
@@ -344,6 +344,7 @@ class _CoverSelectionDialogState extends State<CoverSelectionDialog> {
             source: context.read<CollectionSource>(),
             filters: {filter},
           ),
+          canRemoveFilters: false,
         ),
         fullscreenDialog: true,
       ),