+++ lib/widgets/home_page.dart
@@ -195,10 +195,12 @@ class _HomePageState extends State<HomePage> {
         unawaited(GlobalSearch.registerCallback());
         unawaited(AnalysisService.registerCallback());
         final source = context.read<CollectionSource>();
-        await source.init(
-          loadTopEntriesFirst: settings.homePage == HomePageSetting.collection && settings.homeCustomCollection.isEmpty,
-          canAnalyze: !safeMode,
-        );
+        if (source.initState != SourceInitializationState.full) {
+          await source.init(
+            loadTopEntriesFirst: settings.homePage == HomePageSetting.collection && settings.homeCustomCollection.isEmpty,
+            canAnalyze: !safeMode,
+          );
+        }
       case AppMode.screenSaver:
         final source = context.read<CollectionSource>();
         await source.init(
@@ -221,9 +223,7 @@ class _HomePageState extends State<HomePage> {
       case AppMode.edit:
       case AppMode.setWallpaper:
         await _initViewerEssentials();
-      case AppMode.pickMediaInternal:
-      case AppMode.pickFilterInternal:
-      case AppMode.slideshow:
+      default:
         break;
     }
 
@@ -331,12 +331,7 @@ class _HomePageState extends State<HomePage> {
             );
           },
         );
-      case AppMode.main:
-      case AppMode.pickCollectionFiltersExternal:
-      case AppMode.pickMediaInternal:
-      case AppMode.pickFilterInternal:
-      case AppMode.screenSaver:
-      case AppMode.slideshow:
+      default:
         routeName = _initialRouteName ?? settings.homePage.routeName;
         filters = _initialFilters ?? (settings.homePage == HomePageSetting.collection ? settings.homeCustomCollection : {});
     }